Resolved pull request comments.
