+++ README.md
@@ -73,36 +73,64 @@ git checkout HEAD examples/simple/pi.cc
 
 ### Applying Dredd to a CMake project
 
-The `examples/project` directory contains a small math library to illustrate how to apply dread to a larger project.
-To compile the project, run the following from the root of the repository:
+The `examples/math` directory contains a small math library to illustrate how to apply Dredd to a larger project.
+To compile the math library, run the following from the root of the repository:
 ```
-cd examples/project
+cd examples/math
 mkdir build && cd build
 cmake -G Ninja -DCMAKE_EXPORT_COMPILE_COMMANDS=ON ..
 cmake --build .
 ```
+The `-DCMAKE_EXPORT_COMPILE_COMMANDS=ON` argument causes CMake to output a compilation database called `compile_comands.json` to be output to the `build` directory.
+You can read more about compilation databases for various build systems [here](https://sarcasm.github.io/notes/dev/compilation-database.html).
+
 You can check the output of the tests before mutations have been applied by running 
 `./mathtest/mathtest` from the build directory.
 
-To apply mutants to the program, either do:
+To apply mutants to a single file in the library, run the following command:
 ```
-../../../third_party/clang+llvm-13.0.1/bin/dredd -p compile_commands.js <path-to-file>
+path/to/dredd/repo/third_party/clang+llvm-13.0.1/bin/dredd -p compile_commands.js <path-to-file>
 ``` 
-to mutate a single file. For example, `../../../third_party/clang+llvm-13.0.1/bin/dredd ../math/src/exp.cc` will apply
+For example, `../../../clang+llvm-13.0.1/bin/dredd ../math/src/exp.cc` will apply
 mutants to the file `exp.cc`.
 
-or:
+To view the changes that Dredd has made you can do `git status` to see that `exp.cc` has changed, and `git diff` to see
+the effect that Dredd has had on this file. These changes will be hard to understand as they are not intended to be
+readable by humans.
+
+The `-p` option allows the compilation database generated by CMake above to be passed to Dredd.
+This is so that Dredd knows the correct compiler options to use when processing each source file.
+
+If you wish to mutate multiple source files simultaneously,pass them all to a single invocation of Dredd. 
+For the example project we provide a shell script to do this, `mutate.sh`.
+
+To revert the changes that Dredd made to `exp.cc`, run:
+```
+git checkout HEAD ../math/src/exp.cc
+git status ../math/src/exp.cc
+```
+
+Now, from the projects build directory, run:
 ```
 ../mutate.sh
 ```
-This will run a shell script that mutates all of the `.cc` files in the library.
+This will mutate all of the `.cc` files in the library.
+
+As with the single file above, you can run `git status` to see which files have changed, and `git diff` to see
+the effect that Dredd has had on these files. These changes will be hard to understand as they are not intended to be
+readable by humans.
 
 The project can then be recompiled with
 ```
 cmake --build .
 ```
-Due to the changes made by Dredd, the compile command may lead to warnings being issued but this is expected. Running the test suite with `./mathtest/mathtest` will result in all the tests passing as before. You can specify a
-mutant to apply at runtime by running
+Due to the changes made by Dredd, the compile command may lead to warnings being issued.
+This is expected, because the mutants that Dredd introduces leads to code that violates many good
+programing practices. If your project treats compiler warnings as errors then you will need to disable
+this feature for your projects build system.
+
+Running the test suite with `./mathtest/mathtest` will result in all the tests passing as before, because no mutants are enabled. 
+You can specify a mutant to apply at runtime by setting the `DREDD_ENABLED_MUTATION` environment variable to a non-negative integer value.
 ```
 DREDD_ENABLED_MUTATION=$i$ ./mathtest/mathtest
 ```
@@ -111,10 +139,10 @@ where $i$ is the of the mutation in the modified library generated by Dredd. For
 This may cause some tests to fail or in certain cases, not terminate. There is the possibility of equivalent mutants
 being generated which can lead to all tests still passing.
 
-To clean up the `examples/project` directory, run the following from the build directory:
+To clean up the `examples/math` directory, run the following from the build directory:
 ```
 cd .. && rm -rf build
-git checkout HEAD ../project
+git checkout HEAD .
 ```
 
 ## Building Dredd from source
